backend/requirements.txt
alembic==1.13.3
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
firebase-admin==6.5.0
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==3.0.1
msgpack==1.1.0
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.4
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
SQLAlchemy==2.0.35
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.0.4
